Add the endpoint for the total count getting
