+++ src/VirtoCommerce.DescriptionExportImportModule.Data/Services/ExportPagedDataSourceFactory.cs
@@ -1,4 +1,3 @@
-using VirtoCommerce.CatalogModule.Core.Search;
 using VirtoCommerce.CatalogModule.Core.Services;
 using VirtoCommerce.DescriptionExportImportModule.Core.Models;
 using VirtoCommerce.DescriptionExportImportModule.Core.Services;
@@ -9,22 +8,19 @@ namespace VirtoCommerce.DescriptionExportImportModule.Data.Services
     {
         private readonly IProductEditorialReviewSearchService _productEditorialReviewSearchService;
         private readonly IItemService _itemService;
-        private readonly IListEntrySearchService _listEntrySearchService;
 
         public ExportPagedDataSourceFactory(
             IProductEditorialReviewSearchService productEditorialReviewSearchService,
-            IItemService itemService,
-            IListEntrySearchService listEntrySearchService
-            )
+            IItemService itemService
+        )
         {
             _productEditorialReviewSearchService = productEditorialReviewSearchService;
             _itemService = itemService;
-            _listEntrySearchService = listEntrySearchService;
         }
 
         public IExportPagedDataSource Create(int pageSize, ExportDataRequest request)
         {
-            return new ExportPagedDataSource(_productEditorialReviewSearchService, _itemService, _listEntrySearchService, pageSize, request);
+            return new ExportPagedDataSource(_productEditorialReviewSearchService, _itemService, pageSize, request);
         }
     }
 }
