+++ src/VirtoCommerce.DescriptionExportImportModule.Data/Services/ProductEditorialReviewSearchService.cs
@@ -3,6 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
+using VirtoCommerce.CatalogModule.Core.Model;
+using VirtoCommerce.CatalogModule.Core.Model.Search;
+using VirtoCommerce.CatalogModule.Core.Search;
 using VirtoCommerce.CatalogModule.Data.Model;
 using VirtoCommerce.CatalogModule.Data.Repositories;
 using VirtoCommerce.DescriptionExportImportModule.Core.Models;
@@ -16,15 +19,23 @@ namespace VirtoCommerce.DescriptionExportImportModule.Data.Services
     {
         private readonly Func<ICatalogRepository> _catalogRepositoryFactory;
         private readonly IProductEditorialReviewService _productEditorialReviewService;
+        private readonly IListEntrySearchService _listEntrySearchService;
 
-        public ProductEditorialReviewSearchService(Func<ICatalogRepository> catalogRepositoryFactory, IProductEditorialReviewService productEditorialReviewService)
+        public ProductEditorialReviewSearchService(Func<ICatalogRepository> catalogRepositoryFactory, IProductEditorialReviewService productEditorialReviewService,
+            IListEntrySearchService listEntrySearchService)
         {
             _catalogRepositoryFactory = catalogRepositoryFactory;
             _productEditorialReviewService = productEditorialReviewService;
+            _listEntrySearchService = listEntrySearchService;
         }
 
-        public async Task<ProductEditorialReviewSearchResult> SearchEditorialReviewsAsync(ProductEditorialReviewSearchCriteria criteria)
+        public async Task<ProductEditorialReviewSearchResult> SearchEditorialReviewsAsync(ProductEditorialReviewSearchCriteria criteria, bool deepSearch = false)
         {
+            if (deepSearch)
+            {
+                await ExtendSearchCriteriaForDeepSearch(criteria);
+            }
+
             var result = new ProductEditorialReviewSearchResult();
 
             using var catalogRepository = _catalogRepositoryFactory();
@@ -82,6 +93,96 @@ namespace VirtoCommerce.DescriptionExportImportModule.Data.Services
             }
 
             return sortInfos;
+        }
+
+
+        /// <summary>
+        /// Extend categories criteria with children categories. For the deep searching.
+        /// </summary>
+        /// <param name="searchCriteria"></param>
+        /// <returns></returns>
+        private async Task ExtendSearchCriteriaForDeepSearch(ProductEditorialReviewSearchCriteria searchCriteria)
+        {
+            // All from catalog
+            if (string.IsNullOrEmpty(searchCriteria.Keyword) && searchCriteria.CategoryIds.Length == 0 && searchCriteria.ItemIds.Length == 0)
+            {
+                var listEntrySearchCriteria = new CatalogListEntrySearchCriteria()
+                {
+                    CatalogId = searchCriteria.CatalogId,
+                    SearchInChildren = true,
+                    SearchInVariations = true,
+                    Take = int.MaxValue
+                };
+
+                var listEntrySearchResult = await _listEntrySearchService.SearchAsync(listEntrySearchCriteria);
+
+                if (listEntrySearchResult.TotalCount > 0)
+                {
+                    var listEntries = listEntrySearchResult.Results;
+
+                    var categoriesIds = listEntries.Where(x => x.Type.EqualsInvariant("category")).Select(x => x.Id).ToArray();
+
+                    var productsIds = listEntries.Where(x => x.Type.EqualsInvariant("product")).Select(x => x.Id).ToArray();
+
+                    searchCriteria.CategoryIds = categoriesIds;
+
+                    searchCriteria.ItemIds = productsIds;
+
+                    return;
+                }
+            }
+
+            // All with search by keyword 
+            if (!string.IsNullOrEmpty(searchCriteria.Keyword) && searchCriteria.CategoryIds.Length == 0 && searchCriteria.ItemIds.Length == 0)
+            {
+                var listEntrySearchCriteria = new CatalogListEntrySearchCriteria()
+                {
+                    CatalogId = searchCriteria.CatalogId,
+                    Keyword = searchCriteria.Keyword,
+                    SearchInChildren = true,
+                    SearchInVariations = true,
+                    Take = int.MaxValue
+                };
+
+                var listEntrySearchResult = await _listEntrySearchService.SearchAsync(listEntrySearchCriteria);
+
+                if (listEntrySearchResult.TotalCount > 0)
+                {
+                    var listEntries = listEntrySearchResult.Results;
+
+                    var categoriesIds = listEntries.Where(x => x.Type.EqualsInvariant("category")).Select(x => x.Id).ToArray();
+
+                    var productsIds = listEntries.Where(x => x.Type.EqualsInvariant("product")).Select(x => x.Id).ToArray();
+
+                    searchCriteria.CategoryIds = categoriesIds;
+
+                    searchCriteria.ItemIds = productsIds;
+                }
+            }
+
+            // Extend CategoryIds with children. In case with searching by keyword this will work also. 
+            if (searchCriteria.CategoryIds.Length > 0)
+            {
+                var listEntrySearchCriteria = new CatalogListEntrySearchCriteria()
+                {
+                    CatalogId = searchCriteria.CatalogId,
+                    ObjectType = nameof(Category),
+                    CategoryIds = searchCriteria.CategoryIds,
+                    SearchInChildren = true,
+                    SearchInVariations = true,
+                    Take = int.MaxValue
+                };
+
+                var listEntrySearchResult = await _listEntrySearchService.SearchAsync(listEntrySearchCriteria);
+
+                if (listEntrySearchResult.TotalCount > 0)
+                {
+
+                    var categoryIds = listEntrySearchResult.Results.Select(x => x.Id).ToArray();
+
+                    searchCriteria.CategoryIds = searchCriteria.CategoryIds.Union(categoryIds).ToArray();
+                }
+            }
         }
     }
 }
